tickets/src/providers/StreamingProvider/implementations/listeners/__test__/OrderCreatedListener.test.ts
import { IOrderCreatedEvent, OrderStatus } from '@ictickets/common';
import { Ticket } from '@models/Ticket';
import mongoose from 'mongoose';
import { Message } from 'node-nats-streaming';
import { OrderCreatedListener } from '../OrderCreatedListener';
import NatsProvider from '../../NatsProvider';

const setup = async () => {
  // Create an instance of the listener
  const listener = new OrderCreatedListener(NatsProvider.client);

  // Create and save a ticket
  const ticket = Ticket.build({
    title: 'concert',
    price: 99,
    userId: 'userId',
  });
  await ticket.save();

  // Create the fake data event
  const data: IOrderCreatedEvent['data'] = {
    id: new mongoose.Types.ObjectId().toHexString(),
    version: 0,
    status: OrderStatus.Created,
    userId: 'userId',
    expiresAt: 'expiresAt',
    ticket: {
      id: ticket.id,
      price: ticket.price,
    },
  };

  // @ts-ignore
  const msg: Message = {
    ack: jest.fn(),
  };

  return { listener, ticket, data, msg };
};

it('sets the userId of the ticket', async () => {
  const { listener, ticket, data, msg } = await setup();

  await listener.onMessage(data, msg);

  const updatedTicket = await Ticket.findById(ticket.id);

  expect(updatedTicket!.orderId).toEqual(data.id);
});

it('acks the message', async () => {
  const { listener, data, msg } = await setup();

  await listener.onMessage(data, msg);

  expect(msg.ack).toHaveBeenCalled();
});

it('publishes a ticket updated event', async () => {
  const { listener, data, msg } = await setup();

  await listener.onMessage(data, msg);

  expect(NatsProvider.client.publish).toHaveBeenCalled();

  const ticketUpdatedData = JSON.parse(
    (NatsProvider.client.publish as jest.Mock).mock.calls[0][1],
  );

  expect(data.id).toEqual(ticketUpdatedData.orderId);
});
